refactor(tui): extract message helpers to messages.ts
Move extractResultText and historyToChatMessages out of ChatScreen.tsx
into their own module with co-located tests.
